chore: refactor and use normal snapshots
