+++ src/transform/acorn.ts
@@ -1,41 +1,18 @@
-/**
- * Based on transform.ts but uses oxc-parser and oxc-walker
- */
-import { parseSync } from "oxc-parser";
+import * as acorn from "acorn";
 import MagicString from "magic-string";
-import { walk } from "oxc-walker";
+import { walk } from "estree-walker";
 import type {
   Node,
   CallExpression,
   BlockStatement,
   AwaitExpression,
-} from "oxc-parser";
-
-export interface TransformerOptions {
-  /**
-   * The function names to be transformed.
-   *
-   * @default ['withAsyncContext']
-   */
-  asyncFunctions?: string[];
-  /**
-   * @default 'unctx'
-   */
-  helperModule?: string;
-  /**
-   * @default 'executeAsync'
-   */
-  helperName?: string;
-  /**
-   * Whether to transform properties of an object defined with a helper function. For example,
-   * to transform key `middleware` within the object defined with function `defineMeta`, you would pass:
-   * `{ defineMeta: ['middleware'] }`.
-   * @default {}
-   */
-  objectDefinitions?: Record<string, string[]>;
-}
-
-const kInjected = "__unctx_injected__";
+  Position,
+} from "estree";
+import {
+  type TransformerOptions,
+  defaultTransformerOptions,
+  kInjected,
+} from "./index.js";
 
 type MaybeHandledNode = Node & {
   [kInjected]?: boolean;
@@ -43,10 +20,7 @@ type MaybeHandledNode = Node & {
 
 export function createTransformer(options: TransformerOptions = {}) {
   options = {
-    asyncFunctions: ["withAsyncContext"],
-    helperModule: "unctx",
-    helperName: "executeAsync",
-    objectDefinitions: {},
+    ...defaultTransformerOptions,
     ...options,
   };
 
@@ -66,15 +40,18 @@ export function createTransformer(options: TransformerOptions = {}) {
     if (!options_.force && !shouldTransform(code)) {
       return;
     }
-    const ast = parseSync("", code, {
+    const ast = acorn.parse(code, {
       sourceType: "module",
+      ecmaVersion: "latest",
+      locations: true,
     });
 
     const s = new MagicString(code);
+    const lines = code.split("\n");
 
     let detected = false;
 
-    walk(ast.program, {
+    walk(ast as any, {
       enter(node: Node) {
         if (node.type === "CallExpression") {
           const functionName = _getFunctionName(node.callee);
@@ -82,8 +59,8 @@ export function createTransformer(options: TransformerOptions = {}) {
             transformFunctionArguments(node);
             if (functionName !== "callAsync") {
               const lastArgument = node.arguments[node.arguments.length - 1];
-              if (lastArgument && lastArgument.end) {
-                s.appendRight(lastArgument.end, ",1");
+              if (lastArgument && lastArgument.loc) {
+                s.appendRight(toIndex(lastArgument.loc.end), ",1");
               }
             }
           }
@@ -128,6 +105,10 @@ export function createTransformer(options: TransformerOptions = {}) {
       code: s.toString(),
       magicString: s,
     };
+
+    function toIndex(pos: Position) {
+      return lines.slice(0, pos.line - 1).join("\n").length + pos.column + 1;
+    }
 
     function transformFunctionBody(function_: Node) {
       if (
@@ -175,8 +156,8 @@ export function createTransformer(options: TransformerOptions = {}) {
         },
       });
 
-      if (injectVariable && body.start) {
-        s.appendLeft(body.start + 1, "let __temp, __restore;");
+      if (injectVariable && body.loc) {
+        s.appendLeft(toIndex(body.loc.start) + 1, "let __temp, __restore;");
       }
     }
 
@@ -192,21 +173,21 @@ export function createTransformer(options: TransformerOptions = {}) {
     ) {
       const isStatement = parent?.type === "ExpressionStatement";
 
-      if (!node.start || !node.argument.start) {
+      if (!node.loc || !node.argument.loc) {
         return;
       }
 
-      s.remove(node.start, node.argument.start);
-      s.remove(node.end, node.argument.end);
+      s.remove(toIndex(node.loc.start), toIndex(node.argument.loc.start));
+      s.remove(toIndex(node.loc.end), toIndex(node.argument.loc.end));
 
       s.appendLeft(
-        node.argument.start,
+        toIndex(node.argument.loc.start),
         isStatement
           ? `;(([__temp,__restore]=__executeAsync(()=>`
           : `(([__temp,__restore]=__executeAsync(()=>`,
       );
       s.appendRight(
-        node.argument.end,
+        toIndex(node.argument.loc.end),
         isStatement
           ? `)),await __temp,__restore());`
           : `)),__temp=await __temp,__restore(),__temp)`,
