+++ src/transform/oxc.ts
@@ -1,39 +1,20 @@
-import * as acorn from "acorn";
+/**
+ * Based on transform.ts but uses oxc-parser and oxc-walker
+ */
+import { parseSync } from "oxc-parser";
 import MagicString from "magic-string";
-import { walk } from "estree-walker";
+import { walk } from "oxc-walker";
 import type {
   Node,
   CallExpression,
   BlockStatement,
   AwaitExpression,
-  Position,
-} from "estree";
-
-export interface TransformerOptions {
-  /**
-   * The function names to be transformed.
-   *
-   * @default ['withAsyncContext']
-   */
-  asyncFunctions?: string[];
-  /**
-   * @default 'unctx'
-   */
-  helperModule?: string;
-  /**
-   * @default 'executeAsync'
-   */
-  helperName?: string;
-  /**
-   * Whether to transform properties of an object defined with a helper function. For example,
-   * to transform key `middleware` within the object defined with function `defineMeta`, you would pass:
-   * `{ defineMeta: ['middleware'] }`.
-   * @default {}
-   */
-  objectDefinitions?: Record<string, string[]>;
-}
-
-const kInjected = "__unctx_injected__";
+} from "oxc-parser";
+import {
+  defaultTransformerOptions,
+  kInjected,
+  type TransformerOptions,
+} from "./index.js";
 
 type MaybeHandledNode = Node & {
   [kInjected]?: boolean;
@@ -41,10 +22,7 @@ type MaybeHandledNode = Node & {
 
 export function createTransformer(options: TransformerOptions = {}) {
   options = {
-    asyncFunctions: ["withAsyncContext"],
-    helperModule: "unctx",
-    helperName: "executeAsync",
-    objectDefinitions: {},
+    ...defaultTransformerOptions,
     ...options,
   };
 
@@ -64,18 +42,15 @@ export function createTransformer(options: TransformerOptions = {}) {
     if (!options_.force && !shouldTransform(code)) {
       return;
     }
-    const ast = acorn.parse(code, {
+    const ast = parseSync("", code, {
       sourceType: "module",
-      ecmaVersion: "latest",
-      locations: true,
     });
 
     const s = new MagicString(code);
-    const lines = code.split("\n");
 
     let detected = false;
 
-    walk(ast as any, {
+    walk(ast.program, {
       enter(node: Node) {
         if (node.type === "CallExpression") {
           const functionName = _getFunctionName(node.callee);
@@ -83,8 +58,8 @@ export function createTransformer(options: TransformerOptions = {}) {
             transformFunctionArguments(node);
             if (functionName !== "callAsync") {
               const lastArgument = node.arguments[node.arguments.length - 1];
-              if (lastArgument && lastArgument.loc) {
-                s.appendRight(toIndex(lastArgument.loc.end), ",1");
+              if (lastArgument && lastArgument.end) {
+                s.appendRight(lastArgument.end, ",1");
               }
             }
           }
@@ -129,10 +104,6 @@ export function createTransformer(options: TransformerOptions = {}) {
       code: s.toString(),
       magicString: s,
     };
-
-    function toIndex(pos: Position) {
-      return lines.slice(0, pos.line - 1).join("\n").length + pos.column + 1;
-    }
 
     function transformFunctionBody(function_: Node) {
       if (
@@ -180,8 +151,8 @@ export function createTransformer(options: TransformerOptions = {}) {
         },
       });
 
-      if (injectVariable && body.loc) {
-        s.appendLeft(toIndex(body.loc.start) + 1, "let __temp, __restore;");
+      if (injectVariable && body.start) {
+        s.appendLeft(body.start + 1, "let __temp, __restore;");
       }
     }
 
@@ -197,21 +168,21 @@ export function createTransformer(options: TransformerOptions = {}) {
     ) {
       const isStatement = parent?.type === "ExpressionStatement";
 
-      if (!node.loc || !node.argument.loc) {
+      if (!node.start || !node.argument.start) {
         return;
       }
 
-      s.remove(toIndex(node.loc.start), toIndex(node.argument.loc.start));
-      s.remove(toIndex(node.loc.end), toIndex(node.argument.loc.end));
+      s.remove(node.start, node.argument.start);
+      s.remove(node.end, node.argument.end);
 
       s.appendLeft(
-        toIndex(node.argument.loc.start),
+        node.argument.start,
         isStatement
           ? `;(([__temp,__restore]=__executeAsync(()=>`
           : `(([__temp,__restore]=__executeAsync(()=>`,
       );
       s.appendRight(
-        toIndex(node.argument.loc.end),
+        node.argument.end,
         isStatement
           ? `)),await __temp,__restore());`
           : `)),__temp=await __temp,__restore(),__temp)`,
