+++ test/transform.test.ts
@@ -1,5 +1,5 @@
 import { expect, it, describe } from "vitest";
-import { createTransformer } from "../src/transform";
+import { createTransformer } from "../src/transform/acorn";
 
 describe("transforms", () => {
   const transformer = createTransformer({
@@ -28,15 +28,7 @@ describe("transforms", () => {
         const ctx2 = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        const ctx1 = useSomething()
-        ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-        const ctx2 = useSomething()
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("transforms await as variable", () => {
@@ -48,15 +40,7 @@ describe("transforms", () => {
         const ctx = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        const foo = (([__temp,__restore]=__executeAsync(()=>something())),__temp=await __temp,__restore(),__temp)
-        const bar = hello((([__temp,__restore]=__executeAsync(()=>something())),__temp=await __temp,__restore(),__temp))
-        const ctx = useSomething()
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("transforms await in nested scopes", () => {
@@ -71,18 +55,7 @@ describe("transforms", () => {
         const ctx = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        for (const i of foo) {
-          if (i) {
-            ;(([__temp,__restore]=__executeAsync(()=>i())),await __temp,__restore());
-          }
-        }
-        const ctx = useSomething()
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("transforms await in try-catch", () => {
@@ -101,22 +74,7 @@ describe("transforms", () => {
           return navigateTo('/');
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        let user;
-
-        try {
-          user = (([__temp,__restore]=__executeAsync(()=>fetchUser())),__temp=await __temp,__restore(),__temp);
-        } catch (e) {
-          user = null;
-        }
-
-        if (!user)
-          return navigateTo('/');
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("transforms dot usage", () => {
@@ -128,15 +86,7 @@ describe("transforms", () => {
         const ctx2 = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default ctx.callAsync(async () => {let __temp, __restore;
-        const ctx1 = useSomething()
-        ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-        const ctx2 = useSomething()
-      })
-      "
-    `);
+    ).toMatchSnapshot();
 
     expect(
       transform(`
@@ -146,15 +96,7 @@ describe("transforms", () => {
         const ctx2 = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default x.ctx.callAsync(async () => {let __temp, __restore;
-        const ctx1 = useSomething()
-        ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-        const ctx2 = useSomething()
-      })
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("does not transform non async usage", () => {
@@ -196,19 +138,7 @@ describe("transforms", () => {
         const ctx = useSomething()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-
-        withAsyncContext(async () => {let __temp, __restore;
-          ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-        },1)
-
-        const ctx = useSomething()
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("transforms multiple awaits in same chunk", () => {
@@ -218,13 +148,7 @@ describe("transforms", () => {
         await writeConfig(await readConfig())
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        ;(([__temp,__restore]=__executeAsync(()=>writeConfig((([__temp,__restore]=__executeAsync(()=>readConfig())),__temp=await __temp,__restore(),__temp)))),await __temp,__restore());
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("does not transform non target function", () => {
@@ -266,33 +190,7 @@ describe("transforms", () => {
         }
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default defineSomething({
-        someKey: async () => {let __temp, __restore;
-          const ctx1 = useSomething()
-          ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-          const ctx2 = useSomething()
-        },
-        async someKey () {let __temp, __restore;
-          const ctx1 = useSomething()
-          ;(([__temp,__restore]=__executeAsync(()=>something())),await __temp,__restore());
-          const ctx2 = useSomething()
-        },
-        ...someKey,
-        someKey: 421,
-        someKey () {
-          const ctx1 = useSomething()
-          const ctx2 = useSomething()
-        },
-        async someOtherKey () {
-          const ctx1 = useSomething()
-          await something()
-          const ctx2 = useSomething()
-        }
-      })
-      "
-    `);
+    ).toMatchSnapshot();
   });
 
   it("doesn't transform non-objects", () => {
@@ -309,12 +207,6 @@ describe("transforms", () => {
         if(false) await something()
       })
     `),
-    ).toMatchInlineSnapshot(`
-      "import { executeAsync as __executeAsync } from "unctx";
-      export default withAsyncContext(async () => {let __temp, __restore;
-        if(false) (([__temp,__restore]=__executeAsync(()=>something())),__temp=await __temp,__restore(),__temp)
-      },1)
-      "
-    `);
+    ).toMatchSnapshot();
   });
 });
